Wrote an implementation of SHA256
Wrote an implementation of SHA256, based primarily on psuedocode obtained from wikipedia. It can be improved, and optimised. Fixed a stupid return type mistake in bitwise.h.

+++ include/crypt/bitwise.h
@@ -28,7 +28,7 @@ static uint32_t rotl32(uint32_t rot, uint32_t by)
     return (rot << by) | (rot >> (32 - by));
 }
 
-static uint64_t rotr32(uint32_t rot, uint32_t by)
+static uint32_t rotr32(uint32_t rot, uint32_t by)
 {
     return (rot >> by) | (rot << (32 - by));
 }
